pyproject.toml
[project]
name = "ai-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.62.0",
    "bs4>=0.0.2",
    "duckduckgo-search>=8.1.1",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "wikipedia>=1.4.0",
]
